Moved tests from acp_params.py to test_acp.py Minor fixes in test_alru.py
